style(backend): eslint 규칙 강화에 따른 lint 이슈 일괄 수정

+++ backend/src/common/services/s3.service.ts
@@ -64,7 +64,7 @@ export class S3Service {
     const stream = response.Body as Readable;
     const chunks: Buffer[] = [];
     for await (const chunk of stream) {
-      chunks.push(Buffer.from(chunk));
+      chunks.push(Buffer.from(chunk as Uint8Array));
     }
     return Buffer.concat(chunks);
   }
